add bitstream.fix base warning.

+++ base/TaskQueue.h
@@ -48,7 +48,7 @@ public:
 	}
 
 private:
-	int								_list_size;
+	size_t							_list_size;
 	std::list<T>					_block_queue;
 	std::mutex						_block_mutex;
 	std::condition_variable_any		_empty_notify;
